backend/requirements.txt
python-dotenv
fastapi
supabase
pydantic
uvicorn[standard]
PyJWT
requests